Thief: Ears and Back of the Hair is done.
The back of the hair is finally attached to the body, and was altered a bit to make it easier to add the ponytail. I also added ears (to help me better visualize the hair. The next step will probably be redoing the mouth to make it easier to animate, and then the eyes. Then we work on the rest of the body.