Thief: Hand Model is Done!
The chest looks good enough, and I finished making a model of the hands. I now just need to attach them to the wrists, and then I'm done with them. I may need to make a separate branch in Kraken, to make a quick UV map of her face for an assignment, though.