Thief: Belly is Done, Legs are started!
This model is going along well! The next step is to continue to the ankles (I'll likely stop at the ankles, so that I can shape the shoes, instead of her bare feet). Afterwards, Commit (so that I can save this basic model for adding different clothes on), and then start working on the hands!