Thief: UV Mapping in progress!
The UV Mapping is going along nicely, and I now know how I want to do it. The hair, head, and belt are done. The next step is the feet, and then the hands. After that, it's editing the UV's to fit nicely in their box. (I may separate the pants just so that I can recognize that they're the pants.