Robot Guard: Legs Finished
The basic arms have been shaped, and arms and legs have been mirrored. A head has been created with a core power piece marked on his chest. After getting some fresh eyes on it, the next step will be to shape the chest to apply the circle onto the boxed frame.