Robotic Guard: Tried Adding Bevels
I was able to add bevels in a few places, where it wouldn't edit the placement of the existing UV's, but I don't dare try adding more. If I can figure out how to put the changes into Substance Painter so that I can see how the robot looks, that would be great. Otherwise, this project is officially done.